Refactor SDM, separate AoR, temp update to main
