+++ src/astaire_aor_store.cpp
@@ -29,7 +29,6 @@
 
 #include "log.h"
 #include "utils.h"
-#include "subscriber_data_manager.h"
 #include "astaire_aor_store.h"
 #include "constants.h"
 #include "json_parse_utils.h"
