+++ src/subscriber_data_manager.cpp
@@ -33,6 +33,7 @@ extern "C" {
 #include "log.h"
 #include "utils.h"
 #include "subscriber_data_manager.h"
+#include "astaire_aor_store.h"
 #include "notify_utils.h"
 #include "stack.h"
 #include "pjutils.h"
@@ -43,6 +44,8 @@ extern "C" {
 #include "rapidjson/error/en.h"
 
 /// JSON serialization constants.
+/// These live here, as the core logic of serialization lives in
+/// the AoR::Binding and AoR::Subscription to_json methods.
 static const char* const JSON_BINDINGS = "bindings";
 static const char* const JSON_URI = "uri";
 static const char* const JSON_CID = "cid";
@@ -65,6 +68,7 @@ static const char* const JSON_ROUTES = "routes";
 static const char* const JSON_NOTIFY_CSEQ = "notify_cseq";
 static const char* const JSON_SCSCF_URI = "scscf-uri";
 
+
 /// Helper to delete vectors of bindings safely
 void delete_bindings(ClassifiedBindings& cbs)
 {
@@ -77,16 +81,13 @@ void delete_bindings(ClassifiedBindings& cbs)
 }
 
 /// SubscriberDataManager Methods
-SubscriberDataManager::SubscriberDataManager(Store* data_store,
+SubscriberDataManager::SubscriberDataManager(AoRStore* aor_store,
                                              ChronosConnection* chronos_connection,
                                              AnalyticsLogger* analytics_logger,
                                              bool is_primary) :
   _primary_sdm(is_primary)
 {
-  JsonSerializerDeserializer* serializer = new JsonSerializerDeserializer();
-  std::vector<JsonSerializerDeserializer*> deserializers = { new JsonSerializerDeserializer() };
-
-  _connector = new Connector(data_store, serializer, deserializers);
+  _aor_store = aor_store;
   _chronos_timer_request_sender = new ChronosTimerRequestSender(chronos_connection);
   _notify_sender = new NotifySender();
   _analytics = analytics_logger;
@@ -97,17 +98,17 @@ SubscriberDataManager::~SubscriberDataManager()
 {
   delete _notify_sender;
   delete _chronos_timer_request_sender;
-  delete _connector;
+  delete _aor_store;
 }
 
 /// Retrieve the registration data for a given SIP Address of Record.
 ///
 /// @param aor_id       The SIP Address of Record for the registration
-SubscriberDataManager::AoRPair* SubscriberDataManager::get_aor_data(
+AoRPair* SubscriberDataManager::get_aor_data(
                                           const std::string& aor_id,
                                           SAS::TrailId trail)
 {
-  AoR* aor_data = _connector->get_aor_data(aor_id, trail);
+  AoR* aor_data = _aor_store->get_aor_data(aor_id, trail);
 
   if (aor_data != NULL)
   {
@@ -218,8 +219,8 @@ Store::Status SubscriberDataManager::set_aor_data(
   // how many NOTIFYs we're going to send then we'll have to write back to
   // memcached again
   aor_pair->get_current()->_notify_cseq++;
-  Store::Status rc = _connector->set_aor_data(aor_id,
-                                              aor_pair->get_current(),
+  Store::Status rc = _aor_store->set_aor_data(aor_id,
+                                              aor_pair,
                                               max_expires - now,
                                               trail);
 
@@ -249,7 +250,7 @@ Store::Status SubscriberDataManager::set_aor_data(
 }
 
 void SubscriberDataManager::classify_bindings(const std::string& aor_id,
-                                              SubscriberDataManager::AoRPair* aor_pair,
+                                              AoRPair* aor_pair,
                                               ClassifiedBindings& classified_bindings)
 {
   // We should have been given an empty classified_bindings vector, but clear
@@ -257,7 +258,7 @@ void SubscriberDataManager::classify_bindings(const std::string& aor_id,
   delete_bindings(classified_bindings);
 
   // 1/2: Iterate over original bindings and record those not in current AoR
-  for (std::pair<std::string, SubscriberDataManager::AoR::Binding*> aor_orig_b :
+  for (std::pair<std::string, AoR::Binding*> aor_orig_b :
          aor_pair->get_orig()->bindings())
   {
     if (aor_pair->get_current()->bindings().find(aor_orig_b.first) ==
@@ -273,10 +274,10 @@ void SubscriberDataManager::classify_bindings(const std::string& aor_id,
   }
 
   // 2/2: Iterate over the bindings in the current AoR.
-  for (std::pair<std::string, SubscriberDataManager::AoR::Binding*> aor_current_b :
+  for (std::pair<std::string, AoR::Binding*> aor_current_b :
          aor_pair->get_current()->bindings())
   {
-    SubscriberDataManager::AoR::Bindings::const_iterator aor_orig_b_match =
+    AoR::Bindings::const_iterator aor_orig_b_match =
       aor_pair->get_orig()->bindings().find(aor_current_b.first);
 
     NotifyUtils::ContactEvent event;
@@ -398,12 +399,12 @@ void SubscriberDataManager::expire_subscriptions(AoRPair* aor_pair,
       // The subscription has expired, so remove it. This could be
       // a single one shot subscription though - if so pretend it was
       // part of the original AoR
-      SubscriberDataManager::AoR::Subscriptions::const_iterator aor_orig_s =
+      AoR::Subscriptions::const_iterator aor_orig_s =
         aor_pair->get_orig()->subscriptions().find(i->first);
 
       if (aor_orig_s == aor_pair->get_orig()->subscriptions().end())
       {
-        SubscriberDataManager::AoR::Subscription* s_copy =
+        AoR::Subscription* s_copy =
           aor_pair->get_orig()->get_subscription(i->first);
         *s_copy = *i->second;
       }
@@ -467,764 +468,6 @@ int SubscriberDataManager::expire_bindings(AoR* aor_data,
   return max_expires;
 }
 
-/// SubscriberDataManager::Connector Methods
-
-SubscriberDataManager::Connector::Connector(Store* data_store,
-                               JsonSerializerDeserializer*& serializer,
-                               std::vector<JsonSerializerDeserializer*>& deserializers) :
-  _data_store(data_store),
-  _serializer(serializer),
-  _deserializers(deserializers)
-{
-  // We have taken ownership of the serializer and deserializers.
-  serializer = NULL;
-  deserializers.clear();
-}
-
-SubscriberDataManager::Connector::~Connector()
-{
-  delete _serializer; _serializer = NULL;
-
-  for (JsonSerializerDeserializer* ds : _deserializers)
-  {
-    delete ds; ds = NULL;
-  }
-}
-
-/// Retrieve the registration data for a given SIP Address of Record, creating
-/// an empty record if no data exists for the AoR.
-///
-/// @param aor_id       The SIP Address of Record for the registration
-SubscriberDataManager::AoR* SubscriberDataManager::Connector::get_aor_data(
-                                                 const std::string& aor_id,
-                                                 SAS::TrailId trail)
-{
-  TRC_DEBUG("Get AoR data for %s", aor_id.c_str());
-  AoR* aor_data = NULL;
-
-  std::string data;
-  uint64_t cas;
-  Store::Status status = _data_store->get_data("reg", aor_id, data, cas, trail);
-
-  if (status == Store::Status::OK)
-  {
-    // Retrieved the data, so deserialize it.
-    TRC_DEBUG("Data store returned a record, CAS = %ld", cas);
-    aor_data = deserialize_aor(aor_id, data);
-
-    if (aor_data != NULL)
-    {
-      aor_data->_cas = cas;
-
-      SAS::Event event(trail, SASEvent::REGSTORE_GET_FOUND, 0);
-      event.add_var_param(aor_id);
-      SAS::report_event(event);
-    }
-    else
-    {
-      // Could not deserialize the record. Treat it as not found.
-      TRC_INFO("Failed to deserialize record");
-      SAS::Event event(trail, SASEvent::REGSTORE_DESERIALIZATION_FAILED, 0);
-      event.add_var_param(aor_id);
-      event.add_var_param(data);
-      SAS::report_event(event);
-    }
-  }
-  else if (status == Store::Status::NOT_FOUND)
-  {
-    // Data store didn't find the record, so create a new blank record.
-    aor_data = new AoR(aor_id);
-
-    SAS::Event event(trail, SASEvent::REGSTORE_GET_NEW, 0);
-    event.add_var_param(aor_id);
-    SAS::report_event(event);
-
-    TRC_DEBUG("Data store returned not found, so create new record, CAS = %ld",
-              aor_data->_cas);
-  }
-  else
-  {
-    SAS::Event event(trail, SASEvent::REGSTORE_GET_FAILURE, 0);
-    event.add_var_param(aor_id);
-    SAS::report_event(event);
-  }
-
-  return aor_data;
-}
-
-Store::Status SubscriberDataManager::Connector::set_aor_data(
-                                                const std::string& aor_id,
-                                                AoR* aor_data,
-                                                int expiry,
-                                                SAS::TrailId trail)
-{
-  std::string data = serialize_aor(aor_data);
-
-  SAS::Event event(trail, SASEvent::REGSTORE_SET_START, 0);
-  event.add_var_param(aor_id);
-  SAS::report_event(event);
-
-  Store::Status status = _data_store->set_data("reg",
-                                               aor_id,
-                                               data,
-                                               aor_data->_cas,
-                                               expiry,
-                                               trail);
-
-  TRC_DEBUG("Data store set_data returned %d", status);
-
-  if (status == Store::Status::OK)
-  {
-    SAS::Event event2(trail, SASEvent::REGSTORE_SET_SUCCESS, 0);
-    event2.add_var_param(aor_id);
-    SAS::report_event(event2);
-  }
-  else
-  {
-    SAS::Event event2(trail, SASEvent::REGSTORE_SET_FAILURE, 0);
-    event2.add_var_param(aor_id);
-    SAS::report_event(event2);
-  }
-
-  return status;
-}
-
-/// Serialize the contents of an AoR.
-std::string SubscriberDataManager::Connector::serialize_aor(AoR* aor_data)
-{
-  return _serializer->serialize_aor(aor_data);
-}
-
-/// Deserialize the contents of an AoR
-SubscriberDataManager::AoR* SubscriberDataManager::Connector::deserialize_aor(
-                                                   const std::string& aor_id,
-                                                   const std::string& s)
-{
-  AoR* aor = NULL;
-
-  for (JsonSerializerDeserializer* deserializer : _deserializers)
-  {
-    TRC_DEBUG("Try to deserialize record for %s with JSON deserializer",
-              aor_id.c_str());
-    aor = deserializer->deserialize_aor(aor_id, s);
-
-    if (aor != NULL)
-    {
-      TRC_DEBUG("Deserialization suceeded");
-      return aor;
-    }
-    else
-    {
-      TRC_DEBUG("Deserialization failed");
-    }
-  }
-
-  return aor;
-}
-
-
-/// AoR Methods
-
-/// Default constructor.
-SubscriberDataManager::AoR::AoR(std::string sip_uri) :
-  _notify_cseq(1),
-  _timer_id(""),
-  _scscf_uri(""),
-  _bindings(),
-  _subscriptions(),
-  _cas(0),
-  _uri(sip_uri)
-{
-}
-
-
-/// Destructor.
-SubscriberDataManager::AoR::~AoR()
-{
-  clear(true);
-}
-
-
-/// Copy constructor.
-SubscriberDataManager::AoR::AoR(const AoR& other)
-{
-  common_constructor(other);
-}
-
-// Make sure assignment is deep!
-SubscriberDataManager::AoR& SubscriberDataManager::AoR::operator= (AoR const& other)
-{
-  if (this != &other)
-  {
-    clear(true);
-    common_constructor(other);
-  }
-
-  return *this;
-}
-
-void SubscriberDataManager::AoR::common_constructor(const AoR& other)
-{
-  for (Bindings::const_iterator i = other._bindings.begin();
-       i != other._bindings.end();
-       ++i)
-  {
-    Binding* bb = new Binding(*i->second);
-    _bindings.insert(std::make_pair(i->first, bb));
-  }
-
-  for (Subscriptions::const_iterator i = other._subscriptions.begin();
-       i != other._subscriptions.end();
-       ++i)
-  {
-    Subscription* ss = new Subscription(*i->second);
-    _subscriptions.insert(std::make_pair(i->first, ss));
-  }
-
-  _notify_cseq = other._notify_cseq;
-  _timer_id = other._timer_id;
-  _cas = other._cas;
-  _uri = other._uri;
-  _scscf_uri = other._scscf_uri;
-}
-
-/// Clear all the bindings and subscriptions from this object.
-void SubscriberDataManager::AoR::clear(bool clear_emergency_bindings)
-{
-  for (Bindings::iterator i = _bindings.begin();
-       i != _bindings.end();
-       )
-  {
-    if ((clear_emergency_bindings) || (!i->second->_emergency_registration))
-    {
-      delete i->second;
-      _bindings.erase(i++);
-    }
-    else
-    {
-      ++i;
-    }
-  }
-
-  if (clear_emergency_bindings)
-  {
-    _bindings.clear();
-  }
-
-  for (Subscriptions::iterator i = _subscriptions.begin();
-       i != _subscriptions.end();
-       ++i)
-  {
-    delete i->second;
-  }
-
-  _subscriptions.clear();
-}
-
-
-/// Retrieve a binding by binding identifier, creating an empty one if
-/// necessary.  The created binding is completely empty, even the Contact URI
-/// field.
-SubscriberDataManager::AoR::Binding*
-         SubscriberDataManager::AoR::get_binding(const std::string& binding_id)
-{
-  AoR::Binding* b;
-  AoR::Bindings::const_iterator i = _bindings.find(binding_id);
-  if (i != _bindings.end())
-  {
-    b = i->second;
-  }
-  else
-  {
-    // No existing binding with this id, so create a new one.
-    b = new Binding(_uri);
-    b->_expires = 0;
-    _bindings.insert(std::make_pair(binding_id, b));
-  }
-  return b;
-}
-
-
-/// Removes any binding that had the given ID.  If there is no such binding,
-/// does nothing.
-void SubscriberDataManager::AoR::remove_binding(const std::string& binding_id)
-{
-  AoR::Bindings::iterator i = _bindings.find(binding_id);
-  if (i != _bindings.end())
-  {
-    delete i->second;
-    _bindings.erase(i);
-  }
-}
-
-/// Retrieve a subscription by To tag, creating an empty subscription if
-/// necessary.
-SubscriberDataManager::AoR::Subscription*
-       SubscriberDataManager::AoR::get_subscription(const std::string& to_tag)
-{
-  AoR::Subscription* s;
-  AoR::Subscriptions::const_iterator i = _subscriptions.find(to_tag);
-  if (i != _subscriptions.end())
-  {
-    s = i->second;
-  }
-  else
-  {
-    // No existing subscription with this tag, so create a new one.
-    s = new Subscription;
-    _subscriptions.insert(std::make_pair(to_tag, s));
-  }
-  return s;
-}
-
-
-/// Removes the subscription with the specified tag.  If there is no such
-/// subscription, does nothing.
-void SubscriberDataManager::AoR::remove_subscription(const std::string& to_tag)
-{
-  AoR::Subscriptions::iterator i = _subscriptions.find(to_tag);
-  if (i != _subscriptions.end())
-  {
-    delete i->second;
-    _subscriptions.erase(i);
-  }
-}
-
-/// Remove all the bindings from an AOR object
-void SubscriberDataManager::AoR::clear_bindings()
-{
-  for (Bindings::const_iterator i = _bindings.begin();
-       i != _bindings.end();
-       ++i)
-  {
-    delete i->second;
-  }
-
-  // Clear the bindings map.
-  _bindings.clear();
-}
-
-// Generates the public GRUU for this binding from the address of record and
-// instance-id. Returns NULL if this binding has no valid GRUU.
-pjsip_sip_uri* SubscriberDataManager::AoR::Binding::pub_gruu(pj_pool_t* pool) const
-{
-  pjsip_sip_uri* uri = (pjsip_sip_uri*)PJUtils::uri_from_string(_address_of_record, pool);
-
-  if ((_params.find("+sip.instance") == _params.cend()) ||
-      (uri == NULL) ||
-      !PJSIP_URI_SCHEME_IS_SIP(uri))
-  {
-    // GRUUs are only valid for SIP URIs with an instance-id.
-    return NULL;
-  }
-
-  // The instance parameter might be too short to be a valid GRUU. Specifically
-  // if its less than 2 characters in length, the stripping function will give
-  // us a buffer underrun, so exit now.
-  std::string sip_instance = _params.at("+sip.instance");
-  if (sip_instance.length() < 2)
-  {
-    // instance ID too short to be parsed
-    return NULL;
-  }
-
-  pjsip_param* gr_param = (pjsip_param*) pj_pool_alloc(pool, sizeof(pjsip_param));
-  gr_param->name = STR_GR;
-  pj_strdup2(pool, &gr_param->value, sip_instance.c_str());
-
-  // instance-ids are often of the form '"<urn:..."' - convert that to
-  // just 'urn:...'
-  if (*(gr_param->value.ptr) == '"')
-  {
-    gr_param->value.ptr++;
-    gr_param->value.slen -= 2;
-  }
-
-  if (*(gr_param->value.ptr) == '<')
-  {
-    gr_param->value.ptr++;
-    gr_param->value.slen -= 2;
-  }
-
-  pj_list_push_back((pj_list_type*)&(uri->other_param), (pj_list_type*)gr_param);
-  return uri;
-}
-
-// Utility method to return the public GRUU as a string.
-// Returns "" if this binding has no GRUU.
-std::string SubscriberDataManager::AoR::Binding::pub_gruu_str(pj_pool_t* pool) const
-{
-  pjsip_sip_uri* pub_gruu_uri = pub_gruu(pool);
-
-  if (pub_gruu_uri == NULL)
-  {
-    return "";
-  }
-
-  return PJUtils::uri_to_string(PJSIP_URI_IN_REQ_URI, (pjsip_uri*)pub_gruu_uri);
-}
-
-// Utility method to return the public GRUU surrounded by quotes.
-// Returns "" if this binding has no GRUU.
-std::string SubscriberDataManager::AoR::Binding::pub_gruu_quoted_string(pj_pool_t* pool) const
-{
-  std::string unquoted_pub_gruu = pub_gruu_str(pool);
-
-  if (unquoted_pub_gruu.length() == 0)
-  {
-    return "";
-  }
-
-  std::string ret = "\"" + unquoted_pub_gruu + "\"";
-  return ret;
-}
-
-void SubscriberDataManager::AoR::Binding::
-  to_json(rapidjson::Writer<rapidjson::StringBuffer>& writer) const
-{
-  writer.StartObject();
-  {
-    writer.String(JSON_URI); writer.String(_uri.c_str());
-    writer.String(JSON_CID); writer.String(_cid.c_str());
-    writer.String(JSON_CSEQ); writer.Int(_cseq);
-    writer.String(JSON_EXPIRES); writer.Int(_expires);
-    writer.String(JSON_PRIORITY); writer.Int(_priority);
-
-    writer.String(JSON_PARAMS);
-    writer.StartObject();
-    {
-      for (std::map<std::string, std::string>::const_iterator p = _params.begin();
-           p != _params.end();
-           ++p)
-      {
-        writer.String(p->first.c_str()); writer.String(p->second.c_str());
-      }
-    }
-    writer.EndObject();
-
-    writer.String(JSON_PATH_HEADERS);
-    writer.StartArray();
-    {
-      for (std::list<std::string>::const_iterator p = _path_headers.begin();
-           p != _path_headers.end();
-           ++p)
-      {
-        writer.String(p->c_str());
-      }
-    }
-    writer.EndArray();
-
-    writer.String(JSON_PATHS);
-    writer.StartArray();
-    {
-      for (std::list<std::string>::const_iterator p = _path_uris.begin();
-           p != _path_uris.end();
-           ++p)
-      {
-        writer.String(p->c_str());
-      }
-    }
-    writer.EndArray();
-
-    writer.String(JSON_PRIVATE_ID); writer.String(_private_id.c_str());
-    writer.String(JSON_EMERGENCY_REG); writer.Bool(_emergency_registration);
-  }
-  writer.EndObject();
-}
-
-void SubscriberDataManager::AoR::Binding::from_json(const rapidjson::Value& b_obj)
-{
-
-  JSON_GET_STRING_MEMBER(b_obj, JSON_URI, _uri);
-  JSON_GET_STRING_MEMBER(b_obj, JSON_CID, _cid);
-  JSON_GET_INT_MEMBER(b_obj, JSON_CSEQ, _cseq);
-  JSON_GET_INT_MEMBER(b_obj, JSON_EXPIRES, _expires);
-  JSON_GET_INT_MEMBER(b_obj, JSON_PRIORITY, _priority);
-
-  JSON_ASSERT_CONTAINS(b_obj, JSON_PARAMS);
-  JSON_ASSERT_OBJECT(b_obj[JSON_PARAMS]);
-  const rapidjson::Value& params_obj = b_obj[JSON_PARAMS];
-
-  for (rapidjson::Value::ConstMemberIterator params_it = params_obj.MemberBegin();
-       params_it != params_obj.MemberEnd();
-       ++params_it)
-  {
-    JSON_ASSERT_STRING(params_it->value);
-    _params[params_it->name.GetString()] = params_it->value.GetString();
-  }
-
-  if (b_obj.HasMember(JSON_PATH_HEADERS))
-  {
-    JSON_ASSERT_ARRAY(b_obj[JSON_PATH_HEADERS]);
-    const rapidjson::Value& path_headers_arr = b_obj[JSON_PATH_HEADERS];
-
-    for (rapidjson::Value::ConstValueIterator path_headers_it = path_headers_arr.Begin();
-         path_headers_it != path_headers_arr.End();
-         ++path_headers_it)
-    {
-      JSON_ASSERT_STRING(*path_headers_it);
-      _path_headers.push_back(path_headers_it->GetString());
-    }
-  }
-
-  if (b_obj.HasMember(JSON_PATHS))
-  {
-    JSON_ASSERT_ARRAY(b_obj[JSON_PATHS]);
-    const rapidjson::Value& path_uris_arr = b_obj[JSON_PATHS];
-
-    for (rapidjson::Value::ConstValueIterator path_uris_it = path_uris_arr.Begin();
-         path_uris_it != path_uris_arr.End();
-         ++path_uris_it)
-    {
-      JSON_ASSERT_STRING(*path_uris_it);
-      _path_uris.push_back(path_uris_it->GetString());
-    }
-  }
-
-  JSON_GET_STRING_MEMBER(b_obj, JSON_PRIVATE_ID, _private_id);
-  JSON_GET_BOOL_MEMBER(b_obj, JSON_EMERGENCY_REG, _emergency_registration);
-}
-
-void SubscriberDataManager::AoR::Subscription::
-  to_json(rapidjson::Writer<rapidjson::StringBuffer>& writer) const
-{
-  writer.StartObject();
-  {
-    writer.String(JSON_REQ_URI); writer.String(_req_uri.c_str());
-    writer.String(JSON_FROM_URI); writer.String(_from_uri.c_str());
-    writer.String(JSON_FROM_TAG); writer.String(_from_tag.c_str());
-    writer.String(JSON_TO_URI); writer.String(_to_uri.c_str());
-    writer.String(JSON_TO_TAG); writer.String(_to_tag.c_str());
-    writer.String(JSON_CID); writer.String(_cid.c_str());
-
-    writer.String(JSON_ROUTES);
-    writer.StartArray();
-    {
-      for (std::list<std::string>::const_iterator r = _route_uris.begin();
-           r != _route_uris.end();
-           ++r)
-      {
-        writer.String(r->c_str());
-      }
-    }
-    writer.EndArray();
-
-    writer.String(JSON_EXPIRES); writer.Int(_expires);
-  }
-  writer.EndObject();
-}
-
-void SubscriberDataManager::AoR::Subscription::from_json(const rapidjson::Value& s_obj)
-{
-  JSON_GET_STRING_MEMBER(s_obj, JSON_REQ_URI, _req_uri);
-  JSON_GET_STRING_MEMBER(s_obj, JSON_FROM_URI, _from_uri);
-  JSON_GET_STRING_MEMBER(s_obj, JSON_FROM_TAG, _from_tag);
-  JSON_GET_STRING_MEMBER(s_obj, JSON_TO_URI, _to_uri);
-  JSON_GET_STRING_MEMBER(s_obj, JSON_TO_TAG, _to_tag);
-  JSON_GET_STRING_MEMBER(s_obj, JSON_CID, _cid);
-
-  JSON_ASSERT_CONTAINS(s_obj, JSON_ROUTES);
-  JSON_ASSERT_ARRAY(s_obj[JSON_ROUTES]);
-  const rapidjson::Value& routes_arr = s_obj[JSON_ROUTES];
-
-  for (rapidjson::Value::ConstValueIterator routes_it = routes_arr.Begin();
-       routes_it != routes_arr.End();
-       ++routes_it)
-  {
-    JSON_ASSERT_STRING(*routes_it);
-    _route_uris.push_back(routes_it->GetString());
-  }
-
-  JSON_GET_INT_MEMBER(s_obj, JSON_EXPIRES, _expires);
-}
-
-// Utility function to return the expiry time of the binding or subscription due
-// to expire next. If the function finds no expiry times in the bindings or
-// subscriptions it returns 0. This function should never be called on an empty AoR,
-// so a 0 is indicative of something wrong with the _expires values of AoR members.
-int SubscriberDataManager::AoR::get_next_expires()
-{
-  // Set a temp int to INT_MAX to compare expiry times to.
-  int _next_expires = INT_MAX;
-
-  for (SubscriberDataManager::AoR::Bindings::const_iterator b = _bindings.begin();
-       b != _bindings.end();
-       ++b)
-  {
-    if (b->second->_expires < _next_expires)
-    {
-      _next_expires = b->second->_expires;
-    }
-  }
-  for (SubscriberDataManager::AoR::Subscriptions::const_iterator s = _subscriptions.begin();
-       s != _subscriptions.end();
-       ++s)
-  {
-    if (s->second->_expires < _next_expires)
-    {
-      _next_expires = s->second->_expires;
-    }
-  }
-
-  // If nothing has altered the _next_expires, the AoR is empty and invalid.
-  // Return 0 to indicate there is nothing to expire.
-  if (_next_expires == INT_MAX)
-  {
-    return 0;
-  }
-  // Otherwise we return the value found.
-  return _next_expires;
-}
-
-// Copy all bindings and subscriptions to this AoR
-void SubscriberDataManager::AoR::copy_subscriptions_and_bindings(SubscriberDataManager::AoR* source_aor)
-{
-  for (Bindings::const_iterator i = source_aor->bindings().begin();
-       i != source_aor->bindings().end();
-       ++i)
-  {
-    Binding* src = i->second;
-    Binding* dst = get_binding(i->first);
-    *dst = *src;
-  }
-
-  for (Subscriptions::const_iterator i = source_aor->subscriptions().begin();
-       i != source_aor->subscriptions().end();
-       ++i)
-  {
-    Subscription* src = i->second;
-    Subscription* dst = get_subscription(i->first);
-    *dst = *src;
-  }
-}
-
-//
-// (De)serializer for the JSON SubscriberDataManager format.
-//
-
-SubscriberDataManager::AoR* SubscriberDataManager::JsonSerializerDeserializer::
-  deserialize_aor(const std::string& aor_id, const std::string& s)
-{
-  TRC_DEBUG("Deserialize JSON document: %s", s.c_str());
-
-  rapidjson::Document doc;
-  doc.Parse<0>(s.c_str());
-
-  if (doc.HasParseError())
-  {
-    TRC_DEBUG("Failed to parse document: %s\nError: %s",
-              s.c_str(),
-              rapidjson::GetParseError_En(doc.GetParseError()));
-    return NULL;
-  }
-
-  AoR* aor = new AoR(aor_id);
-
-  try
-  {
-    JSON_ASSERT_CONTAINS(doc, JSON_BINDINGS);
-    JSON_ASSERT_OBJECT(doc[JSON_BINDINGS]);
-    const rapidjson::Value& bindings_obj = doc[JSON_BINDINGS];
-
-    for (rapidjson::Value::ConstMemberIterator bindings_it = bindings_obj.MemberBegin();
-         bindings_it != bindings_obj.MemberEnd();
-         ++bindings_it)
-    {
-      TRC_DEBUG("  Binding: %s", bindings_it->name.GetString());
-      AoR::Binding* b = aor->get_binding(bindings_it->name.GetString());
-
-      JSON_ASSERT_OBJECT(bindings_it->value);
-      const rapidjson::Value& b_obj = bindings_it->value;
-
-      b->from_json(b_obj);
-    }
-
-    JSON_ASSERT_CONTAINS(doc, JSON_SUBSCRIPTIONS);
-    JSON_ASSERT_OBJECT(doc[JSON_SUBSCRIPTIONS]);
-    const rapidjson::Value& subscriptions_obj = doc[JSON_SUBSCRIPTIONS];
-
-    for (rapidjson::Value::ConstMemberIterator subscriptions_it = subscriptions_obj.MemberBegin();
-         subscriptions_it != subscriptions_obj.MemberEnd();
-         ++subscriptions_it)
-    {
-      TRC_DEBUG("  Subscription: %s", subscriptions_it->name.GetString());
-      AoR::Subscription* s = aor->get_subscription(subscriptions_it->name.GetString());
-
-      JSON_ASSERT_OBJECT(subscriptions_it->value);
-      const rapidjson::Value& s_obj = subscriptions_it->value;
-
-      s->from_json(s_obj);
-    }
-
-    JSON_GET_INT_MEMBER(doc, JSON_NOTIFY_CSEQ, aor->_notify_cseq);
-    aor->_timer_id =
-         ((doc.HasMember(JSON_TIMER_ID)) && ((doc[JSON_TIMER_ID]).IsString()) ?
-                                             (doc[JSON_TIMER_ID].GetString()) :
-                                              "");
-    aor->_scscf_uri =
-         ((doc.HasMember(JSON_SCSCF_URI)) && ((doc[JSON_SCSCF_URI]).IsString()) ?
-                                                (doc[JSON_SCSCF_URI].GetString()) :
-                                                 "");
-  }
-  catch(JsonFormatError err)
-  {
-    TRC_INFO("Failed to deserialize JSON document (hit error at %s:%d)",
-             err._file, err._line);
-    delete aor; aor = NULL;
-  }
-
-  return aor;
-}
-
-
-std::string SubscriberDataManager::JsonSerializerDeserializer::serialize_aor(AoR* aor_data)
-{
-  rapidjson::StringBuffer sb;
-  rapidjson::Writer<rapidjson::StringBuffer> writer(sb);
-
-  writer.StartObject();
-  {
-    //
-    // Bindings
-    //
-    writer.String(JSON_BINDINGS);
-    writer.StartObject();
-    {
-      for (AoR::Bindings::const_iterator it = aor_data->bindings().begin();
-           it != aor_data->bindings().end();
-           ++it)
-      {
-        writer.String(it->first.c_str());
-        it->second->to_json(writer);
-      }
-    }
-    writer.EndObject();
-
-    //
-    // Subscriptions.
-    //
-    writer.String(JSON_SUBSCRIPTIONS);
-    writer.StartObject();
-    {
-      for (AoR::Subscriptions::const_iterator it = aor_data->subscriptions().begin();
-           it != aor_data->subscriptions().end();
-           ++it)
-      {
-        writer.String(it->first.c_str());
-        it->second->to_json(writer);
-      }
-    }
-    writer.EndObject();
-
-    // Notify Cseq flag
-    writer.String(JSON_NOTIFY_CSEQ); writer.Int(aor_data->_notify_cseq);
-    writer.String(JSON_TIMER_ID); writer.String(aor_data->_timer_id.c_str());
-    writer.String(JSON_SCSCF_URI); writer.String(aor_data->_scscf_uri.c_str());
-  }
-  writer.EndObject();
-
-  return sb.GetString();
-}
 
 /// ChronosTimerRequestSender Methods
